pyworker: add a basic PathAuditor
Summary:
For now it only checks for symlinks outside of the repo, but can be extended to
check for more.

Caching is used to prevent syscalls to already checked directories, this cache
is intended to be done per-thread, and thus each directory may be checked a
maximum of N, with N being the number of threads. This is deemed reasonable for
now.

Things that the Python pathauditor does that this one doesn't:
 - Case checking for Windows,
 - Verifies no writes to .hg,
 - Windows shortnames aliases,

Reviewed By: quark-zju

Differential Revision: D20265276

fbshipit-source-id: 8e29bedadcaa6df2515e394894ab9e1dd5a9e9a7