connection pool: log remote path
Summary:
We currently log the server we talk to, which is very nice in order to
benchmark performance of Mononoke vs. hg.

Unfortunately, while this allows for breaking down the samples, it's a little
hard to identify the right samples to break down! Historically, we've
approximated this by looking at the client's hostname, but this isn't always
ideal.

Reviewed By: quark-zju

Differential Revision: D20369686

fbshipit-source-id: 785c67fde09e7b7fc4c30121d14f0bc26e92a355