Add configuration setting to repo converter for branch whitelist
Summary: At current, there are some specific AOSP branches that are causing problems for the importer. It would be helpful to be able to run convert jobs that only run on a subset of branches for testing. This commit adds a configuration setting for a whitelist that allows us to specify which repo branches we want to convert, and we only convert those branches.

Reviewed By: tchebb

Differential Revision: D18534728

fbshipit-source-id: bf59605d9c6026e26914c4a7cbe9493bdee77ee5