mononoke: fix integration test regression by reducing number of sqlite shards
Summary:
Even simplest test took ~30 seconds to finish. Turned out most of the time is
sepnt in creating sqlite shards, and the reason we haven't noticed it before is
because we haven't use sqlblob in integration tests until recently (until
February 24 to be precise - D20001261).

We don't need 100 sqlite shards, let's decrease it downto 5.

Reviewed By: HarveyHunt

Differential Revision: D20438107

fbshipit-source-id: e71fd4cabf908e3d92b446fc518a0e5dd64a00bb