hggit: don't write flat map when using indexedlog
Summary:
The flat git map file can be huge and writing to it can hang for many
seconds or minutes on loaded systems since it calls fsync. Since we actually
rely on the indexedlog map file instead of the flat map, let's just stop writing
the flat map when indexedlog is enabled. This was previously done to enable
falling back, but we've had no issues and never needed to fallback.

Reviewed By: quark-zju

Differential Revision: D20443392

fbshipit-source-id: a3e7dbdb2b37e2e99df774f540b05d06414fb279