wait up to 30 seconds for edenfs to exit after SIGKILL
Summary:
When restarting EdenFS processes I've seen a small number of cases where it
took longer than 5 seconds for the kernel to clean up the old EdenFS process
after sending SIGKILL to it on heavily loaded systems.

If we hit this timeout during `edenfsctl restart` then the command exits
unsuccessfully and does not start a new process.  It's therefore desirable to
wait longer if we can.

Reviewed By: wez

Differential Revision: D20484393

fbshipit-source-id: 98a8e5703e7b58c6d809875583347ddb302b9e02