fix signal delivery when using `edenfsctl start --foreground`
Summary:
This fixes handling of Ctrl-C when using `edenfsctl start --foreground` or
`edenfsctl start --gdb`.

The changes in D19861810 to make `exec_daemon()` fork and exec instead of
simply calling exec accidentally broke this.  While it attempted to forward
SIGINT to the child process, in practice this would normally fail in
development environments.  In development environments EdenFS is normally
started using `sudo`, and the CLI does not have permissions to send signals to
this child process since it is running with elevated privileges.  As a result
on Ctrl-C the foreground `edenfsctl` process would die since it did not handle
the error attempting to forward the signal, and the actual EdenFS process
would continue running.

The main goal of D19861810 was to allow telemetry logging of success or
failure on restart.  This changes the `edenfsctl restart` code to use the
existing `start_daemon()` function, and reverts the old behavior of
`exec_daemon()` to use `execve()` to replace the current process.

Reviewed By: genevievehelsel

Differential Revision: D20355252

fbshipit-source-id: 5e61f70e8b411ee023eee6fe06e0cd641f732631