edenapi: remove unused log fields
Summary:
Originally, we wanted to log EdenAPI batch sizes to Scuba in order to A/B test the effect of batch size on fetching speed. It turns out that the original arbitrary choice of 1000 works pretty well, so there isn't really a need to log this to Scuba anymore.

In addition to removing these fields, this diff moves the "http_enabled" column to be logged from the config by the wrapper instead of being manually logged by hg.

Reviewed By: quark-zju

Differential Revision: D20568104

fbshipit-source-id: 1e113489d005e93e6283f7e4a439b1c78b6f0fb9