mononoke: add sampling key to LoggingContainer
Summary:
Add sampling key to LoggingContainer so that we have a way to sample low level blobstore actions for a given high level blobrepo action with the

This is will be used to track the blobs loaded when walking a repo and associate them with paths.

Walker will assign a new sampling_key to each step, and the SamplingHandler attached to SamplingBlobstore will use it correlate the get() with the Node or Path we want to track accesses for. E.g. to build a corpus of blobs per path for compression analysis

Reviewed By: mitrandir77

Differential Revision: D20534844

fbshipit-source-id: 22662b66a57ad4fef044a1108648f4ad8f2dae78