Asyncify stream operations in `statistics_collector`
Summary:
This diff moves all of the Stream operations in `statistics_collector` over
from the old futures api to the new futures-preview api, and in the process
removes all `clone` calls (except where we interface with lower-level
old-futures-based apis).

For reasons that aren't entirely clear to me, rust seems to be unable to
infer the lifetimes of some parameters in `update_statistics`, so I had to
anotate it explicitly.

Reviewed By: krallin

Differential Revision: D20583198

fbshipit-source-id: 1ea247f905649b05ebaabcf27d504d903d5e6365