Asyncify stream internals in `aliasverify`
Summary:
In the previous diffs, I asyncified the api of all `fn`s in
`aliasverify`, but I still had old stream logic, which was hard to
read and required many clones of parameters to handle static lifetime
requirements.

This diff switches all the internal code over to new-style streams.

In the process of doing this, I was able to remove almost all of the
clones, and also switch `get_bounded` from needing to explicitly return
a `Future` via and `async` block to being a standard `async fn` returning
a `Result`.

Reviewed By: krallin

Differential Revision: D20582402

fbshipit-source-id: a98bb31075a41e8d6897b6893cd51ea0c5a94767