Push `compat` down one layer in `backfill_derived_data`
Summary:
In this diff I create a function to handle the logic for choosing a subcommand
to run, returning a Future that we execute. By using `helpers::block_execute`
we're able to extract the spinning up of the fb303 server, which is the key
to making this an async function.

Within each of the subcommands in the new function `choose_subcmd`, this change
also asyncifies the logic inside the `async` block.

Reviewed By: krallin

Differential Revision: D20590742

fbshipit-source-id: a6298e6da3827638cc23369f72ebdd66b5051768